some endian stuff (24 bit visuals)

+++ src/wmbutton.cc
@@ -300,8 +300,8 @@ if (fAction == 0 && icon && showFrameIcon)
 msg("xW/yW: %d/%d, iW/iH: %d/%d", width(), height(), icon->width(), icon->height());
 
 	if (fAction == 0 && icon && showFrameIcon)
-	    g.drawImage(icon, (width() - icon->width()) / 2,
-			      (height() - icon->height()) / 2);
+	    g.drawImage(icon, ((int)width() - (int)icon->width()) / 2,
+			      ((int)height() - (int)icon->height()) / 2);
 
         break;
 #endif
